pyproject.toml
[project]
name = "video-search"
version = "0.1.0"
description = "Local video search webapp with transcript and visual search"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
    "faster-whisper",
    "open-clip-torch",
    "sentence-transformers",
    "chromadb",
    "openai",
    "pillow",
    "watchdog",
    "python-dotenv",
]
